Add minimal test-suite
The ./test script will boot a docker-enabled VM using vagrant,
build the shasoco image and deploy a stack.

It runs some test to make sure the stack is up.

The VM is destroyed after 30 minutes, this let some time to cancel
the deletion, ssh into the VM to run manual tests on the deploy.