Install docker within the shasoco container
Not having docker installed is possible by exposing the docker binary
using a volume. However when running on a host that doesn't share the
same system, I ran into issues with library not being found (apparmor to
name one).

Installing a docker binary in the container seems to solve it, however
it's necessary that it knows to talk with the host docker server... I
hope we won't run into issues.

Anyway, short term goal is to have the infrastructure deployed on 1
server, we can put conditions on this server.