Store domains data in shasocodata container
Containerization is a form of encapsultaion isn't it? And apparently
encapsulation is a good thing.

So by hiding all data managed by the tool from the user, we're
preventing him from trying custom magic with the project configuration
files.

However, we'll need to give him ways to control the configuration too.