Refactor to differenciate deploys and domains
To enforce immutable deployments, it's better to allow multiple deploys
to run in parallel on the same domain. This change is a prelude to the
'metaproxy' task, that'll allow to select the active deploy, and (as a
bonus) to show a maintenance page while backups are performed.

Additionally, we removed the 'deploy' command and replace it with more
atomic operations: `create`, `prepare`, `up`.

The rational behind this is that some failures were hard to recover:
a buggy docker-compose was the end of everything and there was
no way to fix that (except a full redeploy, but this wasn't wanted).
Calling `prepare` on a buggy install again now allows to fix issues.

Third change introduced by this commit is the split-up of docker-compose
definitions per service. Prepare now concatenates each services'
docker-compose definition into a single file. The file tended to grow
out of control.

Close #14, Close #17

Ref #16